renamed: rename all file in capital letter
