+++ ecomerce-app/src/index.css
@@ -1,4 +1,3 @@
-
 @import url("https://grow-n-know.netlify.app/styles/main.css");
 
 body {
