+++ ecomerce-app/src/pages/index.js
@@ -1,9 +1,9 @@
-import {Cart} from "./cart/cart";
-import {Products} from "./products/products";
-import {Wishlisht} from "./wishlisht/wishlisht";
-import {Home} from "./home/home";
-import { Login } from "./auth/login/login";
-import { Signup } from "./auth/signup/signup";
+import {Cart} from "./cart/Cart";
+import {Products} from "./products/Products";
+import {Wishlisht} from "./wishlisht/Wishlisht";
+import {Home} from "./home/Home";
+import { Login } from "./auth/login/Login";
+import { Signup } from "./auth/signup/Signup";
 
 
 export {Cart,Products,Wishlisht,Home,Login,Signup}
